divide cogs
