+++ discordbot.py
@@ -11,7 +11,11 @@ from Repository.PlaylistRepository import IPlaylistRepository
 # init settings
 token = os.environ['DISCORD_BOT_TOKEN']
 # 読み込むコグの名前を格納しておく。
-INITIAL_EXTENSIONS = ['cogs.PlaylistManager']
+INITIAL_EXTENSIONS = [
+    'cogs.PlaylistAdd',
+    'cogs.PlaylistDelete',
+    'cogs.PlaylistOthers'
+]
 
 
 class StartUp:
